write delete get by id and get all test methods and add some logic for not found person in service and controller and write some of that tests

+++ src/main/java/com/example/personcrudapiunitanditegrationtest/controller/PersonController.java
@@ -8,27 +8,40 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.swing.text.html.Option;
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
+
 @RequestMapping("/person")
 @RestController
 
 public class PersonController {
     private final PersonService personService;
+
     @Autowired
     public PersonController(PersonService personService) {
         this.personService = personService;
     }
-
     @GetMapping("/{id}")
-    public ResponseEntity<PersonDto> getById(@PathVariable Long id) throws RuntimeException {
-        return ResponseEntity.status(HttpStatus.ACCEPTED)
-                .body(personService.findById(id));
+    public ResponseEntity<PersonDto> getById(@PathVariable Long id) {
+
+
+            var result = personService.findById(id);
+            if(result == null) return ResponseEntity.notFound().build();
+            else
+                return ResponseEntity.status(HttpStatus.ACCEPTED)
+                        .body(result);
+
     }
+
 
     @GetMapping("/findAll")
     public ResponseEntity<List<PersonDto>> getAllPerson() {
-
-        return ResponseEntity.status(HttpStatus.ACCEPTED).body(personService.findAll());
+        var result = personService.findAll();
+        if (result==null)return ResponseEntity.notFound().build();
+        else
+            return ResponseEntity.status(HttpStatus.ACCEPTED).body(result);
     }
 
     @PostMapping()
@@ -37,12 +50,20 @@ public class PersonController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity update(@PathVariable Long id, @RequestBody PersonDto personDto) throws RuntimeException {
+    public ResponseEntity update(@PathVariable Long id, @RequestBody PersonDto personDto) throws NoSuchElementException {
         return ResponseEntity.status(200).body(personService.update(id, personDto));
     }
 
     @DeleteMapping("/{id}")
-    public void deletePerson(@PathVariable Long id) {
-        personService.deleteById(id);
+    public ResponseEntity<Void> deletePerson(@PathVariable Long id) throws NoSuchElementException {
+        Optional<PersonDto> optionalPersonDto= Optional.ofNullable(personService.findById(id));
+
+            if (optionalPersonDto.isPresent()){
+
+                personService.deleteById(id);
+                return ResponseEntity.noContent().build();
+            }else return ResponseEntity.notFound().build();
+
+
     }
 }
